상단앱바를 구성요소컴포넌트로 분리
